refactor: move study card ordering to study session
