--- src/entities/study-progress/@x/study-session.ts
@@ -1,2 +1,0 @@
-export { buildStudyCardOrder } from "../model/rules";
-export type { CardProgressFields, StudyCardOrderOptions } from "../model/types";
